feat: add experimental runner cli
The context behind this is wanting a standard way to run VS Code tests,
with which we can build an extension that shows VS Code tests in the
test explorer. This does that. The previous APIs are there and will
always be supported as an alternative way to run tests.

See the readme for usage. Integration in js-debug for "debugging" is
not there yet, but it will come.